# src/IO.Swagger/Controllers/ExamApi.cs
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is an opeapi specification for lab project 1  Made by _Oleksandr Myroshnychenko_  Group - 6.04.121.10.19.2
 *
 * OpenAPI spec version: 1.0.11
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ExamApiController : ControllerBase
    { 
        /// <summary>
        /// Returns all exams from db
        /// </summary>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("//api/exam")]
        [ValidateModelState]
        [SwaggerOperation("ApiExamGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Exam>), description: "successful operation")]
        public virtual IActionResult ApiExamGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Exam>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n}, {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Exam>>(exampleJson)
                        : default(List<Exam>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Deletes an exam
        /// </summary>
        /// <remarks>delete an exam</remarks>
        /// <param name="id">Id of an exam that needs to be deleted</param>
        /// <response code="200">Student was successfully deleted</response>
        [HttpDelete]
        [Route("//api/exam/{id}")]
        [ValidateModelState]
        [SwaggerOperation("ApiExamIdDelete")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Exam>), description: "Student was successfully deleted")]
        public virtual IActionResult ApiExamIdDelete([FromRoute][Required]int? id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Exam>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n}, {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Exam>>(exampleJson)
                        : default(List<Exam>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Finds an exam by id
        /// </summary>
        /// <param name="id">ID of an exam that need to return</param>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("//api/exam/{id}")]
        [ValidateModelState]
        [SwaggerOperation("ApiExamIdGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Exam>), description: "successful operation")]
        public virtual IActionResult ApiExamIdGet([FromRoute][Required]int? id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Exam>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n}, {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Exam>>(exampleJson)
                        : default(List<Exam>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Create an exam in the db
        /// </summary>
        /// <param name="entityData">Data that will be added to db as an exam</param>
        /// <response code="200">Entity was successfully added</response>
        [HttpPost]
        [Route("//api/exam")]
        [ValidateModelState]
        [SwaggerOperation("ApiExamPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Exam>), description: "Entity was successfully added")]
        public virtual IActionResult ApiExamPost([FromHeader]EntityData2 entityData)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Exam>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n}, {\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Exam>>(exampleJson)
                        : default(List<Exam>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Update an existing exam
        /// </summary>
        /// <remarks>Updates an existing exam</remarks>
        /// <param name="body">Update an existent exam</param>
        /// <response code="200">successful operation</response>
        [HttpPut]
        [Route("//api/exam")]
        [ValidateModelState]
        [SwaggerOperation("ApiExamPut")]
        [SwaggerResponse(statusCode: 200, type: typeof(Exam), description: "successful operation")]
        public virtual IActionResult ApiExamPut([FromBody]Exam body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Exam));
            string exampleJson = null;
            exampleJson = "{\n  \"Mark\" : 90,\n  \"Id\" : 1,\n  \"SubjectId\" : 1,\n  \"StudentId\" : 1\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Exam>(exampleJson)
                        : default(Exam);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}